Refactor backends to one file per backend
