+++ BackendBench/backends/base.py
@@ -1,0 +1,3 @@
+class Backend:
+    def __init__(self, name):
+        self.name = name
